+++ pkgs/checks/test/extensions/async_test.dart
@@ -16,20 +16,20 @@ void main() {
   group('FutureChecks', () {
     group('completes', () {
       test('succeeds for a future that completes to a value', () async {
-        await checkThat(_futureSuccess()).completes(it()..equals(42));
+        await check(_futureSuccess()).completes(it()..equals(42));
       });
       test('rejects futures which complete as errors', () async {
-        await checkThat(_futureFail()).isRejectedByAsync(
+        await check(_futureFail()).isRejectedByAsync(
           it()..completes(it()..equals(1)),
           actual: ['a future that completes as an error'],
           which: ['threw <UnimplementedError> at:', 'fake trace'],
         );
       });
       test('can be described', () async {
-        await checkThat(it<Future<void>>()..completes())
+        await check(it<Future<void>>()..completes())
             .hasAsyncDescriptionWhich(
                 it()..deepEquals(['  completes to a value']));
-        await checkThat(it<Future<int>>()..completes(it()..equals(42)))
+        await check(it<Future<int>>()..completes(it()..equals(42)))
             .hasAsyncDescriptionWhich(it()
               ..deepEquals([
                 '  completes to a value that:',
@@ -42,11 +42,11 @@ void main() {
       test(
           'succeeds for a future that compeletes to an error of the expected type',
           () async {
-        await checkThat(_futureFail()).throws<UnimplementedError>(
+        await check(_futureFail()).throws<UnimplementedError>(
             it()..has((p0) => p0.message, 'message').isNull());
       });
       test('fails for futures that complete to a value', () async {
-        await checkThat(_futureSuccess()).isRejectedByAsync(
+        await check(_futureSuccess()).isRejectedByAsync(
           it()..throws(),
           actual: ['completed to <42>'],
           which: ['did not throw'],
@@ -54,7 +54,7 @@ void main() {
       });
       test('failes for futures that complete to an error of the wrong type',
           () async {
-        await checkThat(_futureFail()).isRejectedByAsync(
+        await check(_futureFail()).isRejectedByAsync(
           it()..throws<StateError>(),
           actual: ['completed to error <UnimplementedError>'],
           which: [
@@ -64,9 +64,9 @@ void main() {
         );
       });
       test('can be described', () async {
-        await checkThat(it<Future<void>>()..throws()).hasAsyncDescriptionWhich(
+        await check(it<Future<void>>()..throws()).hasAsyncDescriptionWhich(
             it()..deepEquals(['  completes to an error']));
-        await checkThat(it<Future<void>>()..throws<StateError>())
+        await check(it<Future<void>>()..throws<StateError>())
             .hasAsyncDescriptionWhich(it()
               ..deepEquals(['  completes to an error of type StateError']));
       });
@@ -74,19 +74,19 @@ void main() {
 
     group('doesNotComplete', () {
       test('succeeds for a Future that never completes', () async {
-        checkThat(Completer<void>().future).doesNotComplete();
+        check(Completer<void>().future).doesNotComplete();
       });
       test('fails for a Future that completes as a value', () async {
         Object? testFailure;
         runZonedGuarded(() {
           final completer = Completer<String>();
-          checkThat(completer.future).doesNotComplete();
+          check(completer.future).doesNotComplete();
           completer.complete('value');
         }, (e, st) {
           testFailure = e;
         });
         await pumpEventQueue();
-        checkThat(testFailure)
+        check(testFailure)
             .isA<TestFailure>()
             .has((f) => f.message, 'message')
             .isNotNull()
@@ -99,13 +99,13 @@ Actual: a future that completed to 'value\'''');
         Object? testFailure;
         runZonedGuarded(() {
           final completer = Completer<String>();
-          checkThat(completer.future).doesNotComplete();
+          check(completer.future).doesNotComplete();
           completer.completeError('error', StackTrace.fromString('fake trace'));
         }, (e, st) {
           testFailure = e;
         });
         await pumpEventQueue();
-        checkThat(testFailure)
+        check(testFailure)
             .isA<TestFailure>()
             .has((f) => f.message, 'message')
             .isNotNull()
@@ -117,7 +117,7 @@ Which: threw 'error'
 fake trace''');
       });
       test('can be described', () async {
-        await checkThat(it<Future<void>>()..doesNotComplete())
+        await check(it<Future<void>>()..doesNotComplete())
             .hasAsyncDescriptionWhich(
                 it()..deepEquals(['  does not complete']));
       });
@@ -127,26 +127,26 @@ fake trace''');
   group('StreamChecks', () {
     group('emits', () {
       test('succeeds for a stream that emits a value', () async {
-        await checkThat(_countingStream(5)).emits(it()..equals(0));
+        await check(_countingStream(5)).emits(it()..equals(0));
       });
       test('fails for a stream that closes without emitting', () async {
-        await checkThat(_countingStream(0)).isRejectedByAsync(
+        await check(_countingStream(0)).isRejectedByAsync(
           it()..emits(),
           actual: ['a stream'],
           which: ['closed without emitting enough values'],
         );
       });
       test('fails for a stream that emits an error', () async {
-        await checkThat(_countingStream(1, errorAt: 0)).isRejectedByAsync(
+        await check(_countingStream(1, errorAt: 0)).isRejectedByAsync(
           it()..emits(),
           actual: ['a stream with error <UnimplementedError: Error at 1>'],
           which: ['emitted an error instead of a value at:', 'fake trace'],
         );
       });
       test('can be described', () async {
-        await checkThat(it<StreamQueue<void>>()..emits())
+        await check(it<StreamQueue<void>>()..emits())
             .hasAsyncDescriptionWhich(it()..deepEquals(['  emits a value']));
-        await checkThat(it<StreamQueue<int>>()..emits(it()..equals(42)))
+        await check(it<StreamQueue<int>>()..emits(it()..equals(42)))
             .hasAsyncDescriptionWhich(it()
               ..deepEquals([
                 '  emits a value that:',
@@ -156,25 +156,25 @@ fake trace''');
       test('uses a transaction', () async {
         final queue = _countingStream(1, errorAt: 0);
         await softCheckAsync<StreamQueue<int>>(queue, it()..emits());
-        await checkThat(queue).emitsError();
+        await check(queue).emitsError();
       });
     });
 
     group('emitsError', () {
       test('succeeds for a stream that emits an error', () async {
-        await checkThat(_countingStream(1, errorAt: 0))
+        await check(_countingStream(1, errorAt: 0))
             .emitsError<UnimplementedError>();
       });
       test('fails for a stream that closes without emitting an error',
           () async {
-        await checkThat(_countingStream(0)).isRejectedByAsync(
+        await check(_countingStream(0)).isRejectedByAsync(
           it()..emitsError(),
           actual: ['a stream'],
           which: ['closed without emitting an expected error'],
         );
       });
       test('fails for a stream that emits value', () async {
-        await checkThat(_countingStream(1)).isRejectedByAsync(
+        await check(_countingStream(1)).isRejectedByAsync(
           it()..emitsError(),
           actual: ['a stream emitting value <0>'],
           which: ['closed without emitting an error'],
@@ -182,19 +182,19 @@ fake trace''');
       });
       test('fails for a stream that emits an error of the incorrect type',
           () async {
-        await checkThat(_countingStream(1, errorAt: 0)).isRejectedByAsync(
+        await check(_countingStream(1, errorAt: 0)).isRejectedByAsync(
           it()..emitsError<StateError>(),
           actual: ['a stream with error <UnimplementedError: Error at 1>'],
           which: ['emitted an error which is not StateError at:', 'fake trace'],
         );
       });
       test('can be described', () async {
-        await checkThat(it<StreamQueue<void>>()..emitsError())
+        await check(it<StreamQueue<void>>()..emitsError())
             .hasAsyncDescriptionWhich(it()..deepEquals(['  emits an error']));
-        await checkThat(it<StreamQueue<void>>()..emitsError<StateError>())
+        await check(it<StreamQueue<void>>()..emitsError<StateError>())
             .hasAsyncDescriptionWhich(
                 it()..deepEquals(['  emits an error of type StateError']));
-        await checkThat(it<StreamQueue<void>>()
+        await check(it<StreamQueue<void>>()
               ..emitsError<StateError>(
                   it()..has((e) => e.message, 'message').equals('foo')))
             .hasAsyncDescriptionWhich(it()
@@ -207,25 +207,25 @@ fake trace''');
       test('uses a transaction', () async {
         final queue = _countingStream(1);
         await softCheckAsync<StreamQueue<int>>(queue, it()..emitsError());
-        await checkThat(queue).emits((it()..equals(0)));
+        await check(queue).emits((it()..equals(0)));
       });
     });
 
     group('emitsThrough', () {
       test('succeeds for a stream that eventuall emits a matching value',
           () async {
-        await checkThat(_countingStream(5)).emitsThrough(it()..equals(4));
+        await check(_countingStream(5)).emitsThrough(it()..equals(4));
       });
       test('fails for a stream that closes without emitting a matching value',
           () async {
-        await checkThat(_countingStream(4)).isRejectedByAsync(
+        await check(_countingStream(4)).isRejectedByAsync(
           it()..emitsThrough(it()..equals(5)),
           actual: ['a stream'],
           which: ['ended after emitting 4 elements with none matching'],
         );
       });
       test('can be described', () async {
-        await checkThat(it<StreamQueue<int>>()..emitsThrough(it()..equals(42)))
+        await check(it<StreamQueue<int>>()..emitsThrough(it()..equals(42)))
             .hasAsyncDescriptionWhich(it()
               ..deepEquals([
                 '  emits any values then emits a value that:',
@@ -236,25 +236,25 @@ fake trace''');
         final queue = _countingStream(1);
         await softCheckAsync(
             queue, it<StreamQueue<int>>()..emitsThrough(it()..equals(42)));
-        checkThat(queue).emits(it()..equals(0));
+        check(queue).emits(it()..equals(0));
       });
       test('consumes events', () async {
         final queue = _countingStream(3);
-        await checkThat(queue).emitsThrough(it()..equals(1));
-        await checkThat(queue).emits((it()..equals(2)));
+        await check(queue).emitsThrough(it()..equals(1));
+        await check(queue).emits((it()..equals(2)));
       });
     });
 
     group('emitsInOrder', () {
       test('succeeds for happy case', () async {
-        await checkThat(_countingStream(2)).inOrder([
+        await check(_countingStream(2)).inOrder([
           it()..emits(it()..equals(0)),
           it()..emits(it()..equals(1)),
           it()..isDone(),
         ]);
       });
       test('reports which condition failed', () async {
-        await checkThat(_countingStream(1)).isRejectedByAsync(
+        await check(_countingStream(1)).isRejectedByAsync(
           it()..inOrder([it()..emits(), it()..emits()]),
           actual: ['a stream'],
           which: [
@@ -265,7 +265,7 @@ fake trace''');
         );
       });
       test('nestes the report for deep failures', () async {
-        await checkThat(_countingStream(2)).isRejectedByAsync(
+        await check(_countingStream(2)).isRejectedByAsync(
           it()..inOrder([it()..emits(), it()..emits(it()..equals(2))]),
           actual: ['a stream'],
           which: [
@@ -279,7 +279,7 @@ fake trace''');
         );
       });
       test('gets described with the number of conditions', () async {
-        await checkThat(it<StreamQueue<int>>()..inOrder([it(), it()]))
+        await check(it<StreamQueue<int>>()..inOrder([it(), it()]))
             .hasAsyncDescriptionWhich(
                 it()..deepEquals(['  satisfies 2 conditions in order']));
       });
@@ -293,7 +293,7 @@ fake trace''');
                 it()..emits(it()..equals(1)),
                 it()..emits(it()..equals(42)),
               ]));
-        await checkThat(queue).inOrder([
+        await check(queue).inOrder([
           it()..emits(it()..equals(0)),
           it()..emits(it()..equals(1)),
           it()..emits(it()..equals(2)),
@@ -302,8 +302,8 @@ fake trace''');
       });
       test('consumes events', () async {
         final queue = _countingStream(3);
-        await checkThat(queue).inOrder([it()..emits(), it()..emits()]);
-        await checkThat(queue).emits(it()..equals(2));
+        await check(queue).inOrder([it()..emits(), it()..emits()]);
+        await check(queue).emits(it()..equals(2));
       });
     });
 
@@ -311,17 +311,17 @@ fake trace''');
       test(
           'succeeds for a stream that closes without emitting a matching value',
           () async {
-        await checkThat(_countingStream(5)).neverEmits(it()..equals(5));
+        await check(_countingStream(5)).neverEmits(it()..equals(5));
       });
       test('fails for a stream that emits a matching value', () async {
-        await checkThat(_countingStream(6)).isRejectedByAsync(
+        await check(_countingStream(6)).isRejectedByAsync(
           it()..neverEmits(it()..equals(5)),
           actual: ['a stream'],
           which: ['emitted <5>', 'following 5 other items'],
         );
       });
       test('can be described', () async {
-        await checkThat(it<StreamQueue<int>>()..neverEmits(it()..equals(42)))
+        await check(it<StreamQueue<int>>()..neverEmits(it()..equals(42)))
             .hasAsyncDescriptionWhich(it()
               ..deepEquals([
                 '  never emits a value that:',
@@ -332,7 +332,7 @@ fake trace''');
         final queue = _countingStream(2);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..neverEmits(it()..equals(1)));
-        await checkThat(queue).inOrder([
+        await check(queue).inOrder([
           it()..emits(it()..equals(0)),
           it()..emits(it()..equals(1)),
           it()..isDone(),
@@ -342,80 +342,80 @@ fake trace''');
 
     group('mayEmit', () {
       test('succeeds for a stream that emits a matching value', () async {
-        await checkThat(_countingStream(1)).mayEmit(it()..equals(0));
+        await check(_countingStream(1)).mayEmit(it()..equals(0));
       });
       test('succeeds for a stream that emits an error', () async {
-        await checkThat(_countingStream(1, errorAt: 0))
+        await check(_countingStream(1, errorAt: 0))
             .mayEmit(it()..equals(0));
       });
       test('succeeds for a stream that closes', () async {
-        await checkThat(_countingStream(0)).mayEmit(it()..equals(42));
+        await check(_countingStream(0)).mayEmit(it()..equals(42));
       });
       test('consumes a matching event', () async {
         final queue = _countingStream(2);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmit(it()..equals(0)));
-        await checkThat(queue).emits(it()..equals(1));
+        await check(queue).emits(it()..equals(1));
       });
       test('does not consume a non-matching event', () async {
         final queue = _countingStream(2);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmit(it()..equals(1)));
-        await checkThat(queue).emits(it()..equals(0));
+        await check(queue).emits(it()..equals(0));
       });
       test('does not consume an error', () async {
         final queue = _countingStream(1, errorAt: 0);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmit(it()..equals(0)));
-        await checkThat(queue).emitsError<UnimplementedError>(
+        await check(queue).emitsError<UnimplementedError>(
             it()..has((e) => e.message, 'message').equals('Error at 1'));
       });
     });
 
     group('mayEmitMultiple', () {
       test('succeeds for a stream that emits a matching value', () async {
-        await checkThat(_countingStream(1)).mayEmitMultiple(it()..equals(0));
+        await check(_countingStream(1)).mayEmitMultiple(it()..equals(0));
       });
       test('succeeds for a stream that emits an error', () async {
-        await checkThat(_countingStream(1, errorAt: 0))
+        await check(_countingStream(1, errorAt: 0))
             .mayEmitMultiple(it()..equals(0));
       });
       test('succeeds for a stream that closes', () async {
-        await checkThat(_countingStream(0)).mayEmitMultiple(it()..equals(42));
+        await check(_countingStream(0)).mayEmitMultiple(it()..equals(42));
       });
       test('consumes matching events', () async {
         final queue = _countingStream(3);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmitMultiple(it()..isLessThan(2)));
-        await checkThat(queue).emits(it()..equals(2));
+        await check(queue).emits(it()..equals(2));
       });
       test('consumes no events if no events match', () async {
         final queue = _countingStream(2);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmitMultiple(it()..isLessThan(0)));
-        await checkThat(queue).emits(it()..equals(0));
+        await check(queue).emits(it()..equals(0));
       });
       test('does not consume an error', () async {
         final queue = _countingStream(1, errorAt: 0);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmitMultiple(it()..equals(0)));
-        await checkThat(queue).emitsError<UnimplementedError>(
+        await check(queue).emitsError<UnimplementedError>(
             it()..has((e) => e.message, 'message').equals('Error at 1'));
       });
     });
 
     group('isDone', () {
       test('succeeds for an empty stream', () async {
-        await checkThat(_countingStream(0)).isDone();
+        await check(_countingStream(0)).isDone();
       });
       test('fails for a stream that emits a value', () async {
-        await checkThat(_countingStream(1)).isRejectedByAsync(it()..isDone(),
+        await check(_countingStream(1)).isRejectedByAsync(it()..isDone(),
             actual: ['a stream'], which: ['emitted an unexpected value: <0>']);
       });
       test('fails for a stream that emits an error', () async {
         final controller = StreamController<void>();
         controller.addError('sad', StackTrace.fromString('fake trace'));
-        await checkThat(StreamQueue(controller.stream)).isRejectedByAsync(
+        await check(StreamQueue(controller.stream)).isRejectedByAsync(
             it()..isDone(),
             actual: ['a stream'],
             which: ['emitted an unexpected error: \'sad\'', 'fake trace']);
@@ -423,21 +423,21 @@ fake trace''');
       test('uses a transaction', () async {
         final queue = _countingStream(1);
         await softCheckAsync<StreamQueue<int>>(queue, it()..isDone());
-        await checkThat(queue).emits(it()..equals(0));
+        await check(queue).emits(it()..equals(0));
       });
       test('can be described', () async {
-        await checkThat(it<StreamQueue<int>>()..isDone())
+        await check(it<StreamQueue<int>>()..isDone())
             .hasAsyncDescriptionWhich(it()..deepEquals(['  is done']));
       });
     });
 
     group('emitsAnyOf', () {
       test('succeeds for a stream that matches one condition', () async {
-        await checkThat(_countingStream(1)).anyOf(
+        await check(_countingStream(1)).anyOf(
             [it()..emits(it()..equals(42)), it()..emits((it()..equals(0)))]);
       });
       test('fails for a stream that matches no conditions', () async {
-        await checkThat(_countingStream(0)).isRejectedByAsync(
+        await check(_countingStream(0)).isRejectedByAsync(
             it()
               ..anyOf([
                 it()..emits(),
@@ -455,7 +455,7 @@ fake trace''');
             ]);
       });
       test('includes nested details for nested failures', () async {
-        await checkThat(_countingStream(1)).isRejectedByAsync(
+        await check(_countingStream(1)).isRejectedByAsync(
             it()
               ..anyOf([
                 it()..emits(it()..equals(42)),
@@ -475,7 +475,7 @@ fake trace''');
             ]);
       });
       test('gets described with the number of conditions', () async {
-        await checkThat(
+        await check(
                 it<StreamQueue<int>>()..anyOf([it()..emits(), it()..emits()]))
             .hasAsyncDescriptionWhich(
                 it()..deepEquals(['  satisfies any of 2 conditions']));
@@ -489,22 +489,22 @@ fake trace''');
                 it()..emits(it()..equals(10)),
                 it()..emitsThrough(it()..equals(42)),
               ]));
-        await checkThat(queue).emits(it()..equals(0));
+        await check(queue).emits(it()..equals(0));
       });
       test('consumes events', () async {
         final queue = _countingStream(3);
-        await checkThat(queue).anyOf([
+        await check(queue).anyOf([
           it()..emits(it()..equals(1)),
           it()..emitsThrough(it()..equals(1))
         ]);
-        await checkThat(queue).emits(it()..equals(2));
+        await check(queue).emits(it()..equals(2));
       });
     });
   });
 
   group('StreamQueueWrap', () {
     test('can wrap streams in a queue', () async {
-      await checkThat(Stream.value(1)).withQueue.emits();
+      await check(Stream.value(1)).withQueue.emits();
     });
   });
 }