+++ pkgs/checks/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.0
 
 -  **Breaking** `checkThat` renamed to `check`.
+- Added an example.
 
 # 0.1.0