+++ pkgs/checks/CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.1.1-dev
 
-- Added an example.
+-   Added an example.
+-   Include a stack trace in the failure description for unexpected errors from
+    Futures or Streams.
 
 # 0.1.0