+++ pkgs/checks/CHANGELOG.md
@@ -1,7 +1,9 @@
 # 0.2.0
 
--  **Breaking** `checkThat` renamed to `check`.
-- Added an example.
+-   **Breaking** `checkThat` renamed to `check`.
+-   Added an example.
+-   Include a stack trace in the failure description for unexpected errors from
+    Futures or Streams. >>>>>>> master
 
 # 0.1.0