+++ pkgs/checks/CHANGELOG.md
@@ -3,7 +3,7 @@
 -   **Breaking** `checkThat` renamed to `check`.
 -   Added an example.
 -   Include a stack trace in the failure description for unexpected errors from
-    Futures or Streams. >>>>>>> master
+    Futures or Streams.
 
 # 0.1.0