+++ pkgs/checks/CHANGELOG.md
@@ -1,5 +1,6 @@
-# 0.1.1-dev
+# 0.2.0
 
+-   **Breaking** `checkThat` renamed to `check`.
 -   Added an example.
 -   Include a stack trace in the failure description for unexpected errors from
     Futures or Streams.